+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Added `odoo-helper addons uninstall` command
 - Added ability to test odoo-helper-scripts on various debian-based distributions via docker
 - Added `odoo-helper addons list` command, that lists odoo-addons in specified directory
+- Added aliases `odoo-helper flake8` and `odoo-helper pylint`
 - Added automatic configuration checks.
   So, when odoo-helper-scripts provides some new configuration params after update,
   user will be notified about them and asket to update project config file